third_party/WebKit/LayoutTests/animations/custom-properties/registered-var-fallback.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
#target {
  --has-value: 100;
}
</style>
<div id="target"></div>
<script>
for (let name of ['--test', '--has-value']) {
  CSS.registerProperty({
    name,
    syntax: '<number>',
    initialValue: '0',
  });
}

test(() => {
  let animation = target.animate({
    '--test': ['var(--has-value)', 'var(--has-no-value, 200)'],
  }, 100);

  animation.currentTime = 25;
  assert_equals(getComputedStyle(target).getPropertyValue('--test'), '125', 'target at 25%');

  animation.currentTime = 75;
  assert_equals(getComputedStyle(target).getPropertyValue('--test'), '175', 'target at 75%');
}, 'Registered custom property animation keyframes with var() references to unset custom properties should use the fallback value');
</script>
